Rework and CR-D endpoints
